drop all old UI i18n parts from the backend
