+++ frontend/src/global_state/i18n.svelte.ts
@@ -15,7 +15,7 @@ let _i18n: I18n = undefined as any as I18n;
 // Such an API is usually a very bad idea because you can easily mess it up.
 // However, in this case we want it like that on purpose.
 //
-// We only ever set the i18n in the initial routing or inside the <LangSelector>.
+// We only ever set the i18n_email in the initial routing or inside the <LangSelector>.
 // A manual page reload is only necessary in the <LangSelector> though which is a
 // small price when we never need a `get()` in so many other places where we just read
 // the value.
