+++ src/bin/build.rs
@@ -1,11 +1,11 @@
-// use rauthy_models::i18n::ui::I18n;
+// use rauthy_models::i18n_email::ui::I18n;
 // use rauthy_models::language::Language;
 // use std::fs;
 // use std::process::Command;
 // use strum::IntoEnumIterator;
 //
 fn main() {
-    //         let dir = "../../frontend/src/i18n";
+    //         let dir = "../../frontend/src/i18n_email";
     //         println!("Building I18n JSON for UI DEV SSR in {}", dir);
     //
     //         fs::create_dir_all(dir).unwrap();
@@ -18,9 +18,9 @@ fn main() {
     //         .unwrap();
     //
     //         for lang in Language::iter() {
-    //             let i18n = I18n::build(&lang);
+    //             let i18n_email = I18n::build(&lang);
     //             let path = format!("{}/{}.json", dir, lang.as_str());
-    //             let json = serde_json::to_string(&i18n).unwrap();
+    //             let json = serde_json::to_string(&i18n_email).unwrap();
     //             fs::write(path, json).unwrap();
     //         }
     //
