--- src/models/src/i18n/mod.rs
@@ -1,2 +1,0 @@
-pub mod email;
-pub mod ui;
