+++ src/models/src/i18n_email/reset.rs
@@ -1,4 +1,3 @@
-use crate::i18n::ui::SsrJson;
 use crate::language::Language;
 use once_cell::sync::Lazy;
 use serde::Serialize;
@@ -81,18 +80,14 @@ pub struct I18nEmailReset<'a> {
     pub footer: Option<&'a str>,
 }
 
-impl SsrJson for I18nEmailReset<'_> {
-    fn build(lang: &Language) -> Self {
+impl I18nEmailReset<'_> {
+    pub fn build(lang: &Language) -> Self {
         match lang {
             Language::En => Self::build_en(),
             Language::De => Self::build_de(),
             Language::ZhHans => Self::build_zh_hans(),
             Language::Ko => Self::build_ko(),
         }
-    }
-
-    fn as_json(&self) -> String {
-        serde_json::to_string(self).unwrap()
     }
 }
